refactor: move analytics to sub-module
